azure-function/requirements.txt
azure-functions
azure-identity
azure-mgmt-resource
